First Working End-End

+++ src/main/java/com/suresh/projects/movieworld/repositories/MovieRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.suresh.projects.movieworld.entities.Movie;
 
-public interface MovieRepository extends CrudRepository<Movie, Integer> {
+public interface MovieRepository extends CrudRepository<Movie, Long> {
 
 }
